Week6-ML/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
imbalanced-learn
missingno
pyyaml
scipy
fastapi
uvicorn
python-dotenv
joblib
optuna
shap
xgboost
lightgbm